% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_scores.R
\name{get_box_stats}
\alias{get_box_stats}
\title{Parse box score lists}
\usage{
get_box_stats(box_list, team_or_player)
}
\arguments{
\item{box_list}{the list of a match's box score data}

\item{team_or_player}{whether you want box scores for "team" or "player"}
}
\description{
This functions parses scraped box score lists
and returns a data frame of box score statistics for
either the team level, or for players
}
