% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp.R
\name{get_each_pbp}
\alias{get_each_pbp}
\title{Scrape PBP data}
\usage{
get_each_pbp(gamecode, seasoncode)
}
\arguments{
\item{gamecode}{the match number}

\item{seasoncode}{the year the season started with an "E" appended to the front}
}
\description{
This functions scrapes pbp data for a game in a 
season and returns a data frame of pbp data
}
