% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shots.R
\name{get_each_shots}
\alias{get_each_shots}
\title{Scrape shot and shot location data}
\usage{
get_each_shots(gamecode, seasoncode)
}
\arguments{
\item{gamecode}{the match number}

\item{seasoncode}{the year the season started with an "E" appended to the front}
}
\description{
This functions scrapes shooting location data for 
a game in a season and returns a data frame of 
shooting data since the 2015-2016 season
}
