% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/league_meta.R
\name{get_season_rounds}
\alias{get_season_rounds}
\title{Scrape season round information}
\usage{
get_season_rounds(seasoncode)
}
\arguments{
\item{seasoncode}{the year the season started with an "E" appended to the front}
}
\description{
This functions scrapes season rounds data 
and returns a data frame of round information
including round name and number, phase type, 
and round starting and ending dates
}
