% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_results.R
\name{match_results}
\alias{match_results}
\title{Scrape PBP data}
\usage{
match_results(seasoncode, round_phase, round_number)
}
\arguments{
\item{seasoncode}{the year the season started with an "E" appended to the front}

\item{round_phase}{either 'RS' for regular season or 'PO' for playoffs}

\item{round_number}{integer of the round number}
}
\description{
This functions scrapes pbp data for a game in a 
season and returns a data frame of pbp data
}
