% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_h_a_results}
\alias{parse_h_a_results}
\title{Parse Home and Away team metadata}
\usage{
parse_h_a_results(games_list, h_a)
}
\arguments{
\item{games_list}{data frame of a scraped match}

\item{h_a}{whether `"home"` or `"away"`}
}
\description{
This functions parses scraped match results to get metadata
for each team
}
